Rename _distutils_importer to _distutils_hack, as it supplies more than just an importer.
