+++ setuptools/__init__.py
@@ -5,7 +5,7 @@ import functools
 import os
 import re
 
-import _distutils_importer.override  # noqa: F401
+import _distutils_hack.override  # noqa: F401
 
 import distutils.core
 from distutils.errors import DistutilsOptionError
